requirements.txt
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
Django==1.11.13
django-cors-headers==2.2.0
django-webpack-loader==0.6.0
google-api-python-client==1.6.7
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.11.3
idna==2.6
jsonpickle==0.9.6
oauth2client==4.1.2
oauthlib==2.1.0
psycopg2==2.7.4
pyasn1==0.4.3
pyasn1-modules==0.2.1
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
